Simplify annotations in runtime check tests.
It turns out that it is easier to calculate whether or not checks are
needed at the site of the declaration of the interface target, not at
the call site.  Accordingly, it makes sense to put the annotations on
the interface target declarations as well.  This lets us get rid of a
clumsy annotation format that referred to arguments by number, in
favor of simply annotating the formal parameters themselves.

R=scheglov@google.com

Review-Url: https://codereview.chromium.org/3010613003 .