[VM-GC] Add a flag to abort on allocation failure
This is added to help investigate memory usage. By setting an
old-space limit with --old-gen-heap-size and setting the new
--abort-on-oom it is possible to test whether a program can
run in a certain size. Using binary search one can determine
what the minimum size is. Without this flag the VM will slow
down to a crawl when it tries to exceed its heap limit.
R=vegorov@google.com
BUG=

Review-Url: https://codereview.chromium.org/2995063002 .