Better handling of references to `call` in type inference.
During type inference, figuring out the type of a method invocation,
property get, or property set takes three steps:

- findInterfaceMember() looks up the target of the invocation, get, or
  set.

- In the case of a method invocation, additional inference is
  performed (to determine generic parameters).

- getCalleeType() uses the interface member and the receiver type to
  figure out the type of the target.

Previous to this CL, there were two circumstances in which
findInterfaceMember() might return `null`: if the invocation was
dynamic, or the receiver's type was a FunctionType and the method name
is `call`.  In order to allow getCalleeType() to distinguish these two
possibilities, it was necessary to keep track of the method name and
plumb it through to getCalleeType().

That was cumbersome, and it would have interfered with fixing
dartbug.com/30582 (which requires keeping track of the targets of
invocations/gets/sets within complex desugared expressions, in
codepaths that don't have room to easily pass additional information
around).

This CL changes findInterfaceMember to return the sentinel string
`call` to represent the case where the receiver's type is a
FunctionType and the method name is `call`.  This avoids the need for
extra plumbing, since the interface member is already plumbed through
to getCalleeType().

It also does some cleanup of getCalleeType and getSetterType, removing
code branches that have turned out to be unnecessary.

Change-Id: I3295ea139106511005b2067cc9b85b5bc01952c3
Reviewed-on: https://dart-review.googlesource.com/3800
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>