Rework TypeInferenceEngineImpl.computeFormalSafety()
The old technique had the right effect in a few simple corner cases,
but wasn't correct on theoretical grounds.  The new technique is to
keep track, for every method parameter in every interface, the set of
types which might be passed to that method parameter through base
classes, and then compare those types to the declared method parameter
types.

So for instance, in the code:

    class A<T> {
      foo(List<T> argument) {}
    }
    class B extends A<num> {
      foo(List<num> argument) {}
    }

The set of types that might be passed to B.foo's argument is all
subtypes of List<Object> (since a caller might be invoking foo through
the interface A<Object>).  Since List<Object> is not a subtype of
List<num>, B needs a runtime check.

Change-Id: Iae819e9f8c97ec1cf910fbfacf9bc635fccd9706
Reviewed-on: https://dart-review.googlesource.com/4610
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>