Test a corner case of runtime checks (closure called via method interface).
Currently, Dart 2.0 forbids overriding a method with a getter (even if
the types are compatible), so once a method has been torn off, there's
no way to ever call the tear-off through an interface target that is a
method.  However, if we ever relax that rule and allow a method to be
overridden with a getter, it will be possible to call a torn-off
method through an interface target that is a method.  If this ever
happens, we will need to make sure that the proper covariance checks
occur.

Since overriding a method with a getter is currently forbidden, the
new tests verify that the expected compile time error occurs.
However, to future proof us against the possibility that this
restriction is ever relaxed, the tests also contain enough code to
demonstrate the runtime check that would be needed.

Change-Id: Iea3a56d86f661362ff9fb970bcc80c8f80c50f4e
Reviewed-on: https://dart-review.googlesource.com/5322
Reviewed-by: Sigmund Cherem <sigmund@google.com>