[VM, AOT] Fix reaching type of BinaryDoubleOp operand (in strong mode)
This CL adds correction of reaching type of left operand of
BinaryDoubleOp when it is created from devirtualized double operator
(when --experimental-strong-mode is enabled).

The correct type is necessary for code generation of unbox instructions.
The type is not inferred at the first place as VM's type propagator is
not able to infer generic types.

Issue: https://github.com/dart-lang/sdk/issues/30480.
Change-Id: I7795bdd051c568d7cdae3149d51254a4c058d3ba
Reviewed-on: https://dart-review.googlesource.com/7589
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>