Understand _ConstantExpressionError and don't resynthesize corresponding constant expressions.
I think it should not matter whether we resynthesize unresolved AST
or no AST at all in element model. If constant evaluator knows that
no AST means an error.

With summaries we store unresolved AST and detect that something
cannot be resolved during resynthesis. Hm... Maybe we should stop
in this case and return null, as we do now for Kernel.

R=brianwilkerson@google.com, paulberry@google.com

Bug:
Change-Id: Ie1b5dc3b8910e89087f579145337b900a9e481c7
Reviewed-on: https://dart-review.googlesource.com/7637
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>