Handle a forwarding stub corner case.
This CL handles the case where a class overrides a concrete method
with an abstract one, but forwarding stub logic determines that a
concrete method is needed.  In this case we don't need to create a
full forwarding stub; we just need to add an implementation to the
abstract method that already exists.

We annotate this case as `/*@forwardingStub=implementation*/`.

Change-Id: Id0b0c6f84f666161f17942d3b4d7afc115c615f5
Reviewed-on: https://dart-review.googlesource.com/11367
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>