Split finishTopLevel into two phases.
Soon, inference of instance fields will be unified with the logic for
creating forwarding stubs.  When this happens, we will need the order
of inference to be:

1 Finish top level inference of static fields.

2 Finish top level inference of other fields, and create forwarding stubs.

3 Finish top level inference of initializing formals.

Currently, (1) and (3) are performed by the same method,
finishTopLevel().  Splitting it into two methods allows us to put (2)
in between.

Change-Id: I70e1a47e8ad604bdb8f0348c54d6017c0a0b57b5
Reviewed-on: https://dart-review.googlesource.com/12381
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>